pyproject.toml
[project]
name = "ip5-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "frithjofhoppebit",email = "frithjof.hoppe@bit.admin.ch"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "oscal-pydantic-v2 @ git+https://github.com/IP-Cloud-Governance/oscal-pydantic.git@oscal-pydantic-v2",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "datamodel-code-generator (>=0.31.0,<0.32.0)",
    "azure-mgmt-core (>=1.5.0,<2.0.0)",
    "azure-mgmt-policyinsights (>=1.0.0,<2.0.0)",
    "azure-identity (>=1.23.0,<2.0.0)",
    "azure-mgmt-resource (>=23.4.0,<24.0.0)",
]

[tool.poetry]
packages = [{include = "ip5_poc", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
